Initial commit after creating PyCharm project
There is currently one app named 'blog'.  I have
enabled the admin interface by default.

Currently all PyCharm IDE files below ./idea are checked in
*except* workspace.xml and tasks.xml as these contain
user-specific settings.  I think this is fine for now since
currenly I am the only developer working in this project.

If more developers are added to the project we may have
to come up with a different management solution for the
IDE files, since not all developers should be forced to use
the same IDE.