url: add a new /blog/ (index) url
The index page looks exactly like the post_detail page for the latest
blog post, with the exception that a permalink will be added to the
post data when the request url and post url do not match.

In order to add the 'request' variable to the context I had to install
the request template context processor.