tests: add a basic test for comment-form errors
Assert that the preview.html template is loaded.  I went back and made
sure that the form in preview.html was being rendered with the same YUI
classes as the base form.  After doing this the error-case form and
regular-case form render the same.

I also removed the earlier hack to hide the URL form element.  There
must be a better way to accomplish the same thing.