forms: add (optional) to the labels of optional form fields
If a form has mainly required fields (the normal case) then it makes
sense to use (optional) to differentiate the optional fields.
Similarly if most fields are optional then it makes sense to use
(required) to differentiate the required fields.  I plan on using css
to hide one or the other depending on the form.

I thought about using visual cues to differentiate required fields, but
apparently bold text is harder for people to read, and images will
never be as clear as text, even for people without screen readers.