tests/feeds: add tests for my rss and atom feed
In the process I discovered three bugs, two of which were in my own
code and one which was an existing django bug (14656).

The AtomLatestPostsFeed class needs to override get_feed_url(), and the
author_link() and item_author_link() methods were returning relative
urls (I just removed them because they're not really necessary).

The bug in django is that for atom feeds entries do not have a
'published' element.  They have an 'updated' element which was being
populated with the pub-date, but it's not exactly the same thing.