templates: add aria-required attributes using a template tag
Previously I was doing this by adding an attribute to the field in the
Form constructor, but I've decided to move the logic to a template tag
instead.  This gives me greater control in the template itself over
which fields are given aria-required.  It also makes it easy to share
this functionality across different types of forms.