comments: changes to get the new comments app working
At this point the tests all pass, although I suspect that there are
other problems in the comments app waiting to be found.  I had to
rename to comments.py files to comment.py, and for some reason the diff
displayed here doesn't reflect that the files were just renamed.  Hmmm
...