comment moderation: configure email notification
The site managers are always notified when a new comment is posted,
regardless of whether or not it is spam.  If the new comment in not
spam and is a reply to an existing comment the author of the existing
comment is notified, along with any ancestors of the existing comment.
The author of a comment and configure whether or not they want to
recieve emails when someone replies to their comment thread.