templates: use js to make sure the "email me" block is initially checked under certain conditions
If the comment form doesn't have an 'email' field that means that the
email address is implicitly known, and it should be ok to assume that
the user in that case wants to be notified via email.

It's important to only run this java script on the post detail page.
If we run it on the comment preview page it will override any
already-set values for the "email me" box.

This seems to behave ok with page refreshes (ie the initial value isn't
reset) but I've only tested this with firefox.