templates: add an explicit quantity column to the cart table.
I think it's a tiny bit more clear to always explicitly state which
quantity is being used to calculate the row subtotal.  Otherwise if
there is an error when trying to update your cart, the number in the
quantity form field may not be the actual number in the database.  (for
example if the customer tried to add too many items to their cart and
we don't have enough stock to cover it)