templates: improve the look of the update-cart template
Change the button to read "Update Quantity" instead of Update.  I also
removed the Quantity label because IMO it is not necessary on a
one-field form with a clearly labeled button.  I also removed the
initial value from the form - now that there is a quantity column it
just looks confusing.