forms: customize the UserCreationForm
Instead of requiring our customers to remember a username, let them
register with their email address instead.  We will attempt to generate
a unique username from the md5 hash of their email address.  This isn't
foolproof because collisions are still possible, but I thope that they
will be extremely rare.