install django-lazysignup
The lazy signup app comes with it's own set of south migrations, which caused be a little bit of trouble when installing.  The end result was that I had to reset the south tables and clear all pre-existing migrations.  When I re-did the migrations I added the comments app to south.