views: refactor the checkout view
I realized that using a built-in form wizard wasn't going to be flexible enough for my checkout process, so I rolled by own stepper.  There is still a lot of work to do in these views, but it's a start.

The form data isn't actually being saved in the session yet, so as it stands this implementation doesn't do much of anything.