checkout: prevent out-of-order access to checkout urls
Prevent the user from skipping ahead in the checkout process.  They can move back, but cannot move more than 1 step ahead of the last successfully finished step.

If their cart is empty they will always be re-directed to the cart page.

If they haven't officially 'started' the checkout process they will be redirected to the /checkout/ url.  This is the url where a lazy user is created (if necessary) and the data dict is initialized.