checkout: add support for using saved customer-profile data from logged-in users.
If a user is logged in and they have contact information saved to the database, we will attempt to use that information in the checkout contact-info form.  If the data is incomplete, or if they wish to edit said data, the data will be saved to the profile database as soon as the step completes.

I also added a 'cancel' link to the checkout pages to make it easier to cancel an in progress checkout (and thus wipe out all of our stored session data)