checkout: add an order-summary to each page in the checkout process
The order will summarize the number of unique items in the person's cart, the subtotal, and any shipping/tax information that is available.  I removed the 0.00 defaults for many of the currency fields on an order because I need to be able to tell if the tax hasn't been set yet or if it is actually zero.