models: create a new OrderTax model
Instead of storing a single number for the amount of tax applied to an order, associate one or more OrderTax instances with the model.  An OrderTax objects tracks the name of the tax applied (e.g. GST), the tax rate (as a percentage), and the total amount added to the order due to this tax.