templates: add a thin top bar to the top of very page
I'm using a gradient on the top bar to make it look nice.  The exact same gradient is applied to the background html element which gives the impression of the top bar extending across the entire page, making it appear larger and more open.  I've noticed that a number of professional websites use this technique.