views: fix the add_product_count() method to respect product filters
Previously, the add_product_count() method would return the number of products that were in the category, regardless of whether or not the product was in the final product_list.  The new version works, but since it hits the DB multiple times it will be slower.

It would be possible to do this in a single DB query, however not without writing some custom SQL.  This is TBD.