javascript: give each 'view' within a tabview the same height
This height is equal to the maximum height of all the views.  This
is to make the transition between views look a little cleaner in the eyes of the user.  This may become a problem if one of the views is much, much larger than the others, especially if it is not the default view.