models: add an 'unknown' contact method choice
This will be the default when a profile is saved for the first time.  The unknown is to distinguish between the case when a user has never been asked this question, or been asked and selected either 'email' or 'phone'.  The contact-info form doesn't present 'unknown' as a valid choice, of course.