views:  move to a 'ship-to' method for shipping address
First the customer chooses who they would like to ship the order to.  This form is set to automatically submit on change and lives on the same page as the address form.  This is much simpler than the javascript-heavy method I was using easier, and thus will be much easier to modify for non-javascript-enabled users (few as they may be)