models: support associating product images with a product option
At the moment a product image can only be associated with a single product option.  This is just so much simpler than allowing multiple associations, and I think it will suffice for 99% of my products since products with 2+ option categories should be rare, and need to visualize every permeation of those options should be even rarer.