css: add an 'X' ontop of unavailable radio images
The X is built using css before and after objects.  A white opaque X sits underneath a thinner black opaque x.  The effect of this is that I should have an X that can be made out onto of any background colour, to varying degrees.  I'm still not completely happy with the way this looks, but it's better than what I had before.