models: remove the is_active field from Category
I'm not convinced that it's necessary to be able to activate and deactivate categories.  If a category doesn't have any active products in it, it should never be reachable from links on the site.  Viewers can technically type in the url for the category, but no products would show up in it.