models: add an 'order_item' field to WishListItem
If a wish list item is purchased, the purchasing order-item id is linked to the wish list item.  I've linked the order to the wish list item instead of the other way around because with list items can be deleted by the user that created the wish list.  A cascading delete that deleted an order item would be a very bad thing.