javascript: reorganize javascript in templates
While trying to work with global events I realized that for all of my pages I need to have a single <script></script> block.  Currently I was creating a separate <script> block for every page that needed javascript, which I was uneasy about to start with.  The advantage of doing it this way was that pyCharm was able to format my included javascript nicely.

In order to have a single <script> block I decided to move most of my javascript into includable 'js' templates.  Because the file ends with the .js extension pyCharm formats it correctly, but it can still have access to all template variables.