auth: fix a session-loss bug when a lazy user logs in
When a lazy user logs in (to an actual account) there session data is lost because at that point the session-id doesn't match the id of the user to be logged in.  This results in a loss of cart data for logged-in lazy users.  To guard against this manually set the session-id to be that of the to-be-logged-in user in my login routine when the current user is lazy.  Confused, yet?

I noticed that the lazy-signup code does the same thing when it logs in a lazy user.