models: remove the on_mailing_list field from CustomerProfile
In order to keep things simple I'd like the test for whether or not a customer is on the mailing list to be a lookup for their particular email address.  I've removed the on_mailing_list field from all contact info forms.  It's still part of the CreateAccount form, but it's not in the checkout forms.  I'll have to find some other way to ask them to subscribe.